Move TypicalAddressBook to separate file
